# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Civic Cleanup Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpBody"
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/signin:
    post:
      operationId: signIn
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInBody"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /donations:
    get:
      operationId: getDonations
      tags: [donations]
      parameters:
        - name: citizenId
          in: query
          schema:
            type: string
        - name: ngoId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Donations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationsResponse"
    post:
      operationId: createDonation
      tags: [donations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonationBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationResponse"
  /donations/{id}:
    patch:
      operationId: updateDonation
      tags: [donations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDonationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationResponse"

  # COMMUNITY EVENTS
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /messages/{referenceId}:
    get:
      operationId: getMessages
      tags: [messages]
      parameters:
        - name: referenceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesResponse"
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # DUSTBIN
...
  /admin/users/{id}/roles:
    patch:
      operationId: updateUserRoles
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRolesBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /admin/training-modules:
    post:
      operationId: createTrainingModule
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingModuleBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingModuleResponse"
  /admin/redeem-items:
    post:
      operationId: createRedeemItem
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRedeemItemBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedeemItemResponse"
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: ["string", "null"]
        pointsEarned:
          type: ["integer", "null"]
      required: [ok]

    # AUTH
    SignUpBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
      required: [email, password, fullName, role]
    SignInBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
      required: [id, email, fullName]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        token:
          type: string
        roles:
          type: array
          items:
            type: string
        redirectTo:
          type: string
      required: [user, token, roles, redirectTo]
...
    RedeemItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        pointsCost:
          type: integer
        stock:
          type: ["integer", "null"]
        imageEmoji:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]
      required: [id, title, pointsCost]
    RedeemItemResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/RedeemItem"
      required: [data]
...
    TrainingModule:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        lessonCount:
          type: ["integer", "null"]
        sortOrder:
          type: ["integer", "null"]
        requiresPrevious:
          type: ["boolean", "null"]
        createdAt:
          type: string
      required: [id, title, createdAt]
    TrainingModuleResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/TrainingModule"
      required: [data]
...
    CreateTrainingModuleBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        durationMinutes:
          type: integer
        lessonCount:
          type: integer
        sortOrder:
          type: integer
        requiresPrevious:
          type: boolean
      required: [title]

    # SCRAP
...
    Donation:
      type: object
      properties:
        id:
          type: string
        citizenId:
          type: string
        ngoId:
          type: ["string", "null"]
        category:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
        address:
          type: ["string", "null"]
        rewardPoints:
          type: ["integer", "null"]
        scheduledAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        citizenName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, citizenId, category, status, createdAt]
    DonationResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Donation"
      required: [data]
    DonationsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Donation"
      required: [data]
    CreateDonationBody:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        scheduledAt:
          type: string
      required: [category]
    UpdateDonationBody:
      type: object
      properties:
        status:
          type: string
        ngoId:
          type: string
        proofImageUrl:
          type: string
        rewardPoints:
          type: integer

    # EVENTS
...
    Message:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        receiverId:
          type: string
        content:
          type: string
        read:
          type: ["boolean", "null"]
        referenceId:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
        senderAvatar:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, senderId, receiverId, content, createdAt]
    MessageResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Message"
      required: [data]
    MessagesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required: [data]
    SendMessageBody:
      type: object
      properties:
        receiverId:
          type: string
        content:
          type: string
        referenceId:
          type: string
      required: [receiverId, content]

    # DUSTBIN
...
    UpdateUserRolesBody:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
      required: [roles]
    CreateRedeemItemBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        pointsCost:
          type: integer
        stock:
          type: integer
        imageEmoji:
          type: string
        active:
          type: boolean
      required: [title, pointsCost]